Modified the rc property to handle empty channels.
It is possible to not have any valence states in a channel
For example, Fe: 3d64s2, where 3s23p6 remains in the core, has
no explicit l=1 valence states. Before the present modification,
the code would crash because the rc property tried to zip an
empty array. This has been corrected.