Spurious blank line removed at end of input file.
The last line should be "0", indicating to ATOMPAW
that it is time to shut down. Before this modification,
an extra "\n" introduced a blank line in the input file
before the final "0", which lead ATOMPAW to quit
with an error message. Now, with this spurious blank line
removed, the code quits elegantly.

This has no consequence on the functionality of the code,
as this "problem" only occured when ATOMPAW was about
to quit anyways.